Added TextControl to unify code of TextEdit and TextLine
